🚚 rename SDK file
