+++ sdk/nexent/core/agents/__init__.py
@@ -1,4 +1,4 @@
-from .code_agent import CoreAgent
-from .agent_const import ModelConfig, ToolConfig, AgentConfig, AgentRunInfo, AgentHistory
+from .core_agent import CoreAgent
+from .agent_model import ModelConfig, ToolConfig, AgentConfig, AgentRunInfo, AgentHistory
 
 __all__ = ["CoreAgent", "ModelConfig", "ToolConfig", "AgentConfig", "AgentRunInfo", "AgentHistory"]
