+++ sdk/nexent/core/agents/nexent_agent.py
@@ -1,10 +1,12 @@
 from typing import List
 
 from threading import Event
-from .agent_const import ModelConfig, ToolConfig, AgentConfig
+from smolagents import ActionStep, TaskStep, AgentText, handle_agent_output_types
+from ..utils.observer import ProcessType
+from .agent_model import ModelConfig, ToolConfig, AgentConfig, AgentHistory
 from ..utils.observer import MessageObserver
 from ..models.openai_llm import OpenAIModel
-from .code_agent import CoreAgent
+from .core_agent import CoreAgent
 from ..tools import (
     EXASearchTool,
     KnowledgeBaseSearchTool,
@@ -13,7 +15,7 @@ from ..tools import (
     GetEmailTool)           # Used for tool creation, do not delete!!!
 
 
-class AgentCreateFactory:
+class NexentAgent:
     def __init__(self, observer: MessageObserver,
                  model_config_list: List[ModelConfig],
                  stop_event: Event,
@@ -33,6 +35,8 @@ class AgentCreateFactory:
         self.model_config_list = model_config_list
         self.stop_event = stop_event
         self.mcp_tool_collection = mcp_tool_collection
+
+        self.agent = None
 
 
     def create_model(self, model_cite_name: str):
@@ -109,7 +113,7 @@ class AgentCreateFactory:
                 raise ValueError(f"Error in creating tool: {e}")
 
             try:
-                managed_agents_list = [self.create_single_agent(agent_config) for agent_config in agent_config.managed_agents]
+                managed_agents_list = [self.create_single_agent(sub_agent_config) for sub_agent_config in agent_config.managed_agents]
             except Exception as e:
                 raise ValueError(f"Error in creating managed agent: {e}")
 
@@ -126,6 +130,71 @@ class AgentCreateFactory:
                 managed_agents=managed_agents_list
             )
             agent.stop_event = self.stop_event
+
             return agent
         except Exception as e:
             raise ValueError(f"Error in creating agent, agent name: {agent_config.name}, Error: {e}")
+
+    def add_history_to_agent(self, history: List[AgentHistory]):
+        """
+        Add conversation history to agent's memory
+
+        Args:
+            history: List of conversation messages with role and content
+        """
+        if history is None:
+            return
+
+        if not isinstance(self.agent, CoreAgent):
+            raise TypeError(f"agent must be a CoreAgent object, not {type(self.agent)}")
+
+        if not all(isinstance(msg, AgentHistory) for msg in history):
+            raise TypeError("history must be a list of AgentHistory objects")
+
+        self.agent.memory.reset()
+        # Add conversation history to memory sequentially
+        for msg in history:
+            if msg.role == 'user':
+                # Create task step for user message
+                self.agent.memory.steps.append(TaskStep(task=msg.content))
+            elif msg.role == 'assistant':
+                self.agent.memory.steps.append(ActionStep(action_output=msg.content, model_output=msg.content))
+
+    def agent_run_with_observer(self, query: str, reset=True):
+        if not isinstance(self.agent, CoreAgent):
+            raise TypeError(f"agent must be a CoreAgent object, not {type(self.agent)}")
+
+        observer = self.agent.observer
+        try:
+            for step_log in self.agent.run(query, stream=True, reset=reset):
+                # Add content to observer
+                if not isinstance(step_log, ActionStep):
+                    continue
+                # Keep duration
+                if hasattr(step_log, "duration"):
+                    observer.add_message("", ProcessType.TOKEN_COUNT, str(round(float(step_log.duration), 2)))
+
+                if hasattr(step_log, "error") and step_log.error is not None:
+                    observer.add_message("", ProcessType.ERROR, str(step_log.error))
+
+            final_answer = step_log  # Last log is the run's final_answer
+            final_answer = handle_agent_output_types(final_answer)
+
+            if isinstance(final_answer, AgentText):
+                observer.add_message(self.agent.agent_name, ProcessType.FINAL_ANSWER, final_answer.to_string())
+            else:
+                observer.add_message(self.agent.agent_name, ProcessType.FINAL_ANSWER, str(final_answer))
+
+            # Check if we need to stop from external stop_event
+            if self.agent.stop_event.is_set():
+                observer.add_message(self.agent.agent_name, ProcessType.ERROR,
+                                     "Agent execution interrupted by external stop signal")
+        except Exception as e:
+            observer.add_message(agent_name=self.agent.agent_name, process_type=ProcessType.ERROR,
+                                 content=f"Error in interaction: {str(e)}")
+            raise ValueError(f"Error in interaction: {str(e)}")
+
+    def set_agent(self, agent: CoreAgent):
+        if not isinstance(agent, CoreAgent):
+            raise TypeError(f"agent must be a CoreAgent object, not {type(agent)}")
+        self.agent = agent
